return to ichar_traits (needs debugging..)
need a different word order. not the one induced by the character order.

consider aA aB Ac in the dictionary. that must be the ascending order in
insensitive mode. in sensitive mode, we may want to add AB. that must go
between aA and aB, so it can be located in both sensitive and
insensitive mode...